Add tests for reads
Treat tempo as a single register. Read tests implemented, writes to
follow. Took a while to work out that the process needed to be in the
background. Also had to adapt the test script to run a TCP server to
allow the control node to have access to specific sessions.