Fix bug with comms, implement write test
Initially I was using the `(c/exec)` form to try to execute commands on
the nodes but this requires an ssh context which isn't (obviously)
available as part of the nemesis process.

As a fallback I opted for shelling out to netcat (`nc`) with the `-q 1`
option to close one second after stdin is read, otherwise the process
hangs.