Basic test now working - singlethreaded, 2s gap
Jepsen (the test framework) is more often used on programs with some
strong consistency guarantees. Link only claims to be eventually
consistent with the last write winning for state changes like tempo or
"playing?".

To get a test with a linearizable history, even in the single threaded
mode, requires an initial write to overcome an issue on the Jepsen side.
Tests start with nil by default whereas the Link library has a notion of
default tempo (120). This meant that if the first operation in a test
was a read then it would fail with `nil != 120`.

The next issue was to separate the operations by 2 seconds. This is
because the current test looks for linearizability rather than eventual
consistency. The Link library prefers local updates within a 1 second
window - note to self: might this cause failures of linearizability
tests? If so, why? - so a 2 second gap between operations is used to
work around this.