WIP: extract and graph offset measurements
Attempting to plot offset measurements relative to the leader in the
session. At the moment all nodes are within 1ms of each other but with a
4ms offset from the leader. This may be some kind of constant relating
to Docker networking.

The information would probably work just as well on two separate plots
which is an option if I can't get this to work.