Show tempo divergence with graph regions
As a measurement of effectiveness, the amount of time in a session that
Link manages to keep all nodes converging on a tempo and beat seems
valid.

Turned out to be a fairly tricky analysis to write. It relies on the
fact that all nodes report their statuses "on the beat", so grouping
these where the second timings match, the tempi match and all nodes 1 to
5 are present gives an indication of when agreement is reached.

Further thoughts - divergence is affected by the number of consecutive
writes. To maintain fair tests under different network conditions we
should standardize on the order and frequency of read and write ops.

To allow for reproducibility and error detection, all data used to
produce the graph output should also be written to disk.

Graph layout - is is possible to make it clearer when a nemesis is
running vs when tempo has diverged?