Calculate and show convergence time
This commit works out the amount of time that all nodes agree on a
tempo. The initialization period (where nodes are yet to join the
session) is ignored, as is the final period of divergence if a node
changes tempo after the others have left.

The observations from the logs aren't perfect so this required some
smoothing to avoid false positives e.g. where we didn't observe all 5
nodes in the same time observation but the tempo remained constant
across the neighbouring observations