Graph all Link sessions and timing offsets
Instead of graphing only the offset measurements for the winning
session, the script now graphs all sessions. This gives a clearer
picture of how offset timings are calculated and also an indication of
when a node leaves the main session due to network partitions.

Displaying special characters in gnuplot is tricky so some characters
are substituted by a hypen (`-`) in the graph key.