Create partition scheme for line graph
Creating a line partition seemed worthwhile as a way to stress the
terminal nodes under Link's use of reliable broadcast.

This was initially hampered by a lack of documentation around the
format that (nemesis/partitioner) expects. This turns out to be

```
{n2 #{n1 n4}, n5 #{n1 n3}, n4 #{n2 n3}, n1 #{n2 n5}, n3 #{n5 n4}}
```

a map where the keys are strings (node names) and the values are sets of
nodes that they are connected to. The example above shows a ring graph.

Once this was established, the implementation was straightforward.