Add per-node packet logging using iptables
This commit adds an operation to the start of the test which looks up
all IP addresses for nodes n1 to n5 (currently hard coded) and creates a
pass through rule in IP tables on each node, monitoring the UDP packets
received from them.

In conjunction with the `iptables -vL` command, this gives a fairly
accurate view of how much data is transferred as part of a test run.

This data is then ready to be parsed into tables for output later.