Add options for selecting delay distribution via CLI
Adds options for `normal`, (gaussian) `constant` and `pareto`
distributions which are passed to the NetEm command being run on each
node.

The leader (n1) is *not* slowed down. Symmetrical delays (on both sender
and receiver) are impossible to measure accurately with Link's current
synchronization algorithm as they cancel each other out.